# BAU: Speed up headings spec