### Jira link

BAU

### What?

Small test change to speed things up.  

Before:
```
Top 10 slowest examples (62.08 seconds, 34.5% of total time):
  JS behaviour render table tools on the top and bottom
    11.94 seconds ./spec/features/headings_spec.rb:4
  JS behaviour is able to open close specific headings
    6.95 seconds ./spec/features/headings_spec.rb:25
```

After:

```
Top 2 slowest examples (4.02 seconds, 99.7% of total time):
  JS behaviour is able to open close specific headings
    2.01 seconds ./spec/features/headings_spec.rb:25
  JS behaviour render table tools on the top and bottom
    2 seconds ./spec/features/headings_spec.rb:4
```

### Why?

This makes the test suite about 15% faster.